Fix fp16 LoRA unscale crash after validation in remaining DreamBooth LoRA scripts
Follow-up to #13895, which fixed this for examples/dreambooth/train_dreambooth_lora.py.
The same fp16 footgun is present in the other DreamBooth LoRA training scripts: under
`--mixed_precision="fp16"`, `cast_training_params(..., dtype=torch.float32)` keeps the
trainable LoRA params in fp32, but `log_validation` rebuilds the in-loop validation
pipeline around the *live* training transformer (`transformer=unwrap_model(transformer)`)
and then casts it to fp16. That downcasts the fp32 LoRA params, so the next optimizer
step raises `ValueError: Attempting to unscale FP16 gradients`.

Apply the same fix across the remaining scripts:
- flux, flux_kontext, qwen_image, hidream, and advanced flux: drop `dtype=torch_dtype`
  from `pipeline.to(accelerator.device, ...)` (keep the device move), matching #13895.
- z_image and the flux2 variants: the cast is `pipeline.to(dtype=torch_dtype)` with no
  device move, immediately followed by `enable_model_cpu_offload()`, so just drop the
  cast line. Frozen weights already use `weight_dtype` and the offload call handles
  device placement.

The final (post-training) validation in every script builds a fresh pipeline from the
saved weights, so it is unaffected either way.

+++ examples/dreambooth/train_dreambooth_lora_flux_kontext.py
@@ -200,7 +200,7 @@ def log_validation(
         f"Running validation... \n Generating {args.num_validation_images} images with prompt:"
         f" {args.validation_prompt}."
     )
-    pipeline = pipeline.to(accelerator.device, dtype=torch_dtype)
+    pipeline = pipeline.to(accelerator.device)
     pipeline.set_progress_bar_config(disable=True)
     pipeline_args_cp = pipeline_args.copy()
